Moved all code into a single file petl_ldap3.py and modified the setup.py to suit.
Modified the _iter_ldap_query function to better "stream" results as per the PETL norm.
Previously the function would perform a search and then continue paging the search until all results were collected
and then it would iterate those results and yield each row. This was fine for small result sets but could pose an issue for very large result sets.
Now the _iter_ldap_query function will perform a search for a single page of data (default 500 results), which it will then iterate over.
At the end of the results page the sever response will be checked for a paging cookie. If it exists a new search will be performed for the next page and the process repeats.
In this way hopefully no more than the page_size number of results will ever be in memory at once.

The results handling has also been updated and should better handle cases where the LDAP server returns single values in lists.
For cases where the attribute is a multi valued attribute the code will now use the new parameter joiner and join all of the values together into a string.